pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
# See https://www.python.org/dev/peps/pep-0621/
name = "biosimdb-app"
dynamic = ["version"]  # read from biosimdb_app/__init__.py
description = "An application for the BioSimDB project to store and download biomolecular simulation data."
authors = [{name = "Jas Kalayan", email = "jas.kalayan@stfc.ac.uk"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: Flask"
]
keywords = ["BioSimDB", "database", "biomolecular simulation", 
    "molecular dynamics"]
requires-python = ">=3.8"
dependencies = [
    "Flask==2.2.3",
    "Flask-MySQLdb==1.0.1",
    "boto3==1.34.54",
    "aiida-core>=2.4.0,<3",
    "Werkzeug==2.2.2",
    "gunicorn==21.2.0", 
    "pyvis==0.3.2",
    "python-dotenv==1.0.1",
    "MDAnalysis==2.7.0",
]

[project.urls]
Source = "https://github.com/PSDI-UK/biosimdb-app"

[project.optional-dependencies]
testing = [
    "pgtest==1.3.2",
    "wheel==0.43.0",
    "coverage[toml]",
    "pytest==8.2.2",
    "pytest-cov==5.0.0",
    "pytest-sugar==1.0.0"
]
pre-commit = [
    "pre-commit==3.7.1",
    "pylint==3.2.5"
]